+++ wgpu-core/src/command/ray_tracing.rs
@@ -22,7 +22,9 @@ use crate::{
         TlasBuildEntry, TlasInstance, TlasPackage, TraceBlasBuildEntry, TraceBlasGeometries,
         TraceBlasTriangleGeometry, TraceTlasInstance, TraceTlasPackage,
     },
-    resource::{AccelerationStructure, Blas, Buffer, Labeled, StagingBuffer, Tlas},
+    resource::{
+        AccelerationStructure, Blas, BlasCompactState, Buffer, Labeled, StagingBuffer, Tlas,
+    },
     scratch::ScratchBuffer,
     snatch::SnatchGuard,
     track::PendingTransition,
@@ -335,10 +337,16 @@ impl Global {
 
         let cmd_buf_raw = cmd_buf_data.encoder.open()?;
 
+        let mut blas_s_compactable = Vec::new();
         let mut descriptors = Vec::new();
 
         for storage in &blas_storage {
-            descriptors.push(map_blas(storage, scratch_buffer.raw(), &snatch_guard)?);
+            descriptors.push(map_blas(
+                storage,
+                scratch_buffer.raw(),
+                &snatch_guard,
+                &mut blas_s_compactable,
+            )?);
         }
 
         build_blas(
@@ -348,6 +356,7 @@ impl Global {
             input_barriers,
             &descriptors,
             scratch_buffer_barrier,
+            blas_s_compactable,
         );
 
         if tlas_present {
@@ -664,10 +673,16 @@ impl Global {
 
         let cmd_buf_raw = cmd_buf_data.encoder.open()?;
 
+        let mut blas_s_compactable = Vec::new();
         let mut descriptors = Vec::new();
 
         for storage in &blas_storage {
-            descriptors.push(map_blas(storage, scratch_buffer.raw(), &snatch_guard)?);
+            descriptors.push(map_blas(
+                storage,
+                scratch_buffer.raw(),
+                &snatch_guard,
+                &mut blas_s_compactable,
+            )?);
         }
 
         build_blas(
@@ -677,6 +692,7 @@ impl Global {
             input_barriers,
             &descriptors,
             scratch_buffer_barrier,
+            blas_s_compactable,
         );
 
         if tlas_present {
@@ -794,6 +810,15 @@ impl CommandBufferMutable {
 
                     command_index_guard.next_acceleration_structure_build_command_index += 1;
                     for blas in build.blas_s_built.iter() {
+                        let mut state_lock = blas.compacted_state.lock();
+                        *state_lock = match *state_lock {
+                            BlasCompactState::Compacted => {
+                                unreachable!("Should be validated out in build.")
+                            }
+                            // Reset the compacted state to idle. This means any prepares, before mapping their
+                            // internal buffer, will terminate.
+                            _ => BlasCompactState::Idle,
+                        };
                         *blas.built_index.write() = Some(build_command_index);
                     }
 
@@ -1215,6 +1240,10 @@ fn map_blas<'a>(
     storage: &'a BlasStore<'_>,
     scratch_buffer: &'a dyn hal::DynBuffer,
     snatch_guard: &'a SnatchGuard,
+    blases_compactable: &mut Vec<(
+        &'a dyn hal::DynBuffer,
+        &'a dyn hal::DynAccelerationStructure,
+    )>,
 ) -> Result<
     hal::BuildAccelerationStructureDescriptor<
         'a,
@@ -1231,12 +1260,27 @@ fn map_blas<'a>(
     if blas.update_mode == wgt::AccelerationStructureUpdateMode::PreferUpdate {
         log::info!("only rebuild implemented")
     }
+    let raw = blas.try_raw(snatch_guard)?;
+
+    let state_lock = blas.compacted_state.lock();
+    if let BlasCompactState::Compacted = *state_lock {
+        return Err(BuildAccelerationStructureError::CompactedBlas(
+            blas.error_ident(),
+        ));
+    }
+
+    if blas
+        .flags
+        .contains(wgpu_types::AccelerationStructureFlags::ALLOW_COMPACTION)
+    {
+        blases_compactable.push((blas.compaction_buffer.as_ref().unwrap().as_ref(), raw));
+    }
     Ok(hal::BuildAccelerationStructureDescriptor {
         entries,
         mode: hal::AccelerationStructureBuildMode::Build,
         flags: blas.flags,
         source_acceleration_structure: None,
-        destination_acceleration_structure: blas.try_raw(snatch_guard)?,
+        destination_acceleration_structure: raw,
         scratch_buffer,
         scratch_buffer_offset: *scratch_buffer_offset,
     })
@@ -1253,6 +1297,10 @@ fn build_blas<'a>(
         dyn hal::DynAccelerationStructure,
     >],
     scratch_buffer_barrier: hal::BufferBarrier<dyn hal::DynBuffer>,
+    blas_s_for_compaction: Vec<(
+        &'a dyn hal::DynBuffer,
+        &'a dyn hal::DynAccelerationStructure,
+    )>,
 ) {
     unsafe {
         cmd_buf_raw.transition_buffers(&input_barriers);
@@ -1279,6 +1327,20 @@ fn build_blas<'a>(
 
     let mut source_usage = hal::AccelerationStructureUses::empty();
     let mut destination_usage = hal::AccelerationStructureUses::empty();
+    for &(buf, blas) in blas_s_for_compaction.iter() {
+        unsafe {
+            cmd_buf_raw.transition_buffers(&[hal::BufferBarrier {
+                buffer: buf,
+                usage: hal::StateTransition {
+                    from: BufferUses::ACCELERATION_STRUCTURE_QUERY,
+                    to: BufferUses::ACCELERATION_STRUCTURE_QUERY,
+                },
+            }])
+        }
+        unsafe { cmd_buf_raw.read_acceleration_structure_compact_size(blas, buf) }
+        destination_usage |= hal::AccelerationStructureUses::COPY_SRC;
+    }
+
     if blas_present {
         source_usage |= hal::AccelerationStructureUses::BUILD_OUTPUT;
         destination_usage |= hal::AccelerationStructureUses::BUILD_INPUT