+++ wgpu/src/api/blas.rs
@@ -1,6 +1,6 @@
 use alloc::{boxed::Box, vec::Vec};
 
-use wgt::WasmNotSendSync;
+use wgt::{WasmNotSend, WasmNotSendSync};
 
 use crate::dispatch;
 use crate::{Buffer, Label};
@@ -215,4 +215,50 @@ pub struct ContextBlasBuildEntry<'a> {
     pub(crate) blas: &'a dispatch::DispatchBlas,
     #[expect(dead_code)]
     pub(crate) geometries: ContextBlasGeometries<'a>,
+}
+
+/// Error occurred when trying to asynchronously prepare a blas for compaction.
+#[derive(Clone, PartialEq, Eq, Debug)]
+pub struct BlasAsyncError;
+static_assertions::assert_impl_all!(BlasAsyncError: Send, Sync);
+
+impl core::fmt::Display for BlasAsyncError {
+    fn fmt(&self, f: &mut core::fmt::Formatter<'_>) -> core::fmt::Result {
+        write!(
+            f,
+            "Error occurred when trying to asynchronously prepare a blas for compaction"
+        )
+    }
+}
+
+impl core::error::Error for BlasAsyncError {}
+
+impl Blas {
+    /// Asynchronously prepares this BLAS for compaction. The callback is called once all builds
+    /// using this BLAS are finished and the BLAS is compactable. This can be checked using
+    /// [`Blas::ready_for_compaction`]. Rebuilding this BLAS will reset its compacted state, and it
+    /// will need to be prepared again.
+    ///
+    /// ### Interaction with other functions
+    /// On native, `queue.submit(..)` and polling devices (that is calling `instance.poll_all` or
+    /// `device.poll`) with [`PollType::Poll`] may call the callback. On native, polling devices with
+    /// [`PollType::Wait`] (or [`PollType::WaitForSubmissionIndex`] with a submission index greater
+    /// than the last submit the BLAS was used in) will guarantee callback is called.
+    ///
+    /// [`PollType::Poll`]: wgpu_types::PollType::Poll
+    /// [`PollType::Wait`]: wgpu_types::PollType::Wait
+    /// [`PollType::WaitForSubmissionIndex`]: wgpu_types::PollType::WaitForSubmissionIndex
+    pub fn prepare_compaction_async(
+        &self,
+        callback: impl FnOnce(Result<(), BlasAsyncError>) + WasmNotSend + 'static,
+    ) {
+        self.inner.prepare_compact_async(Box::new(callback));
+    }
+
+    /// Checks whether this BLAS is ready for compaction. The returned value is `true` if
+    /// [`Blas::prepare_compaction_async`]'s callback was called with a non-error value, otherwise
+    /// this is `false`.
+    pub fn ready_for_compaction(&self) -> bool {
+        self.inner.ready_for_compaction()
+    }
 }